Improved module architecture
